requirements.txt
dotenv==0.9.9
python3-discogs-client==2.8
requests==2.32.5
discordwebhook==1.0.3